requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-anymail==6.1.0
idna==2.8
Pillow==6.1.0
pycrypto==2.6.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sparkpost==1.3.6
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
